Export getDefaultKeyBinding and KeyBindingUtil
In [Advanced Topics: Key
Bindings](https://facebook.github.io/draft-js/docs/advanced-topics-key-bindings.html)
there is an
[example](https://facebook.github.io/draft-js/docs/advanced-topics-key-bindings.html#example)
where both `getDefaultKeyBinding()` and `KeyBindingUtil` are used, but
neither are exported by `draft-js`. This PR exports both
`getDefaultKeyBinding()` and `KeyBindingUtil` from `draft-js`.

Also in the above example, `getDefaultKeyBinding()` was not imported.
This PR adds `getDefaultKeyBinding()` to the imports for the example.