Exclude trailing block for type toggling if at zero offset
Summary:Resolves #86.

In cases where the user triple-clicks on a block, then toggles the block type, it feels weird to have the trailing block also change block type.

The `SelectionState` resulting from the triple-click is correct, but to improve the interaction here, let's discard the trailing block for this modifier.
Closes https://github.com/facebook/draft-js/pull/111

Reviewed By: spicyj

Differential Revision: D2987166

fb-gh-sync-id: 66dea913955ac5eef354f3222ac5afd7e6fdb92d
shipit-source-id: 66dea913955ac5eef354f3222ac5afd7e6fdb92d