Fix linter warnings
Summary:Fixes linter warnings.

Also adds linting to the CI build, but warnings still exit with `0` which doesn't cause the suite to fail. I'm not even sure whether you guys want the suite to fail on this condition; though the contributing docs asks users to make sure that the linter passes before submitting a PR.
Closes https://github.com/facebook/draft-js/pull/78

Reviewed By: spicyj

Differential Revision: D2987193

fb-gh-sync-id: bfcf661be729fd9f27c65a8237973ce6ff25cb16
shipit-source-id: bfcf661be729fd9f27c65a8237973ce6ff25cb16