Move critical editor style into JS
There have been a number of bug reports caused by Draft.css not being included. The main one is that the editor can start flipping out if `white-space: pre-wrap` is not applied.

Users should generally be including Draft.css, and should be aware of behavior gaps if they choose not to. Documenting this accordingly.