removes iOS format tooltips everywhere
fixes #214

thanks to @bradleyayers excellent digging it turns out there's a
CSS property you can assign to disable the bold/italics/underline
for iOS. although you can techincally apply this style seperately
, since the editor breaks due to the unexpected dom mutations i
think it's best if this actually gets defaultly added to the root
of the editor. which this commit does.