Add table styles to documentation
We have a PR to add a new page to the docs,[1] and it includes a table
showing the mapping of block types to elements. However, we don't have
good styles for tables in the documentation.

[1]: https://github.com/facebook/draft-js/pull/387

Adding these styles should unblock that PR so we can merge it and update
the Draft.js website.