Add default text direction
See #634 for discussion.

By default, Draft.js sets direction to LTR when text has no direction. In
an RTL environment, it means that empty editors show the cursor on the
wrong side, until the user starts typing some RTL text.

The new defaultDirection property of EditorState is used for that. It is sent
to UnicodeBidiService as default direction.

Closes #634.