support `editorKey` prop on <Editor />
allows callers of <Editor /> to pass a custom `editorKey` to a draft
instance. This fixes universal rendering mismatches between client and
the server.

Instead of generating a random key when the component initializes, it
waits for a custom one. If one is not passed, its key is empty until it
renders and then it receives a random one when mounted if one was not
passed in as a prop.