Make contentState.createFromBlockArray work with old and new API
As part of the changing API the signature of
contentState.createFromBlockArray changed.

At some point (recently I think) this was changed to no longer support
the old API, and we didn't quite have the 0.9.1 style examples in place,
so it wasn't caught.

This adjusts the 'createFromBlockArray' to work with both the old and
new APIs.