Use DraftEntity type instead of EntityMap type in helpers
While supporting both the old and new APIs in Draft.js for Entity
manipulation, we are calling to the old DraftEntity module under the
hood instead of using the new system that would use an Immutable.js
OrderedMap called the EntityMap.

The DraftEntity module now has private methods for entity manipulation
that don't warn, and the public API which we are deprecating has
warnings and then calls the private methods.

In short, there are places where the flow typing wasn't right because
it was expecting EntityMap, and not DraftEntity.

I don't know why flow didn't catch this when we merged
https://github.com/facebook/draft-js/pull/814

test plan: `flow src`