Fix flow error caused by passing command to RichUtils.handleKeyCommand
The 'handleKeyCommand' takes either a 'DraftEditorCommand' type or any
string.  This allows users to have their own custom editor command
types. I'm not sure how common this is, but our current flow typing
allows it.

The flow typing of 'RichUtils.handleKeyCommand' is more strict - it only
allows 'DraftEditorCommand' types, and not custom strings. This causes
an issue in one of our use cases, where we pass the 'command' received
in 'handleKeyCommand' through to 'RichUtils.handleKeyCommand'.

It seems like the 'RichUtils.handleKeyCommand' should have an API
similar to 'DraftEditor.handleKeyCommand', and so we're loosening the
flow type on 'RichTuils.handleKeyCommand'.

Test Plan: `flow`