Change underscore method names to bypass lint about private methods
Private methods should not be called as public methods. A recent PR
(https://github.com/facebook/draft-js/pull/814) added warnings to the
public-but-deprecated API of DraftEntity, and then moved the logic for
those methods into private methods that can be called within the
Draft.js framework.

These methods are "private" in the sense that we are only calling them
within Draft. They should not be used outside of the Draft library,
because they will be removed in an upcoming version of the library.

We have a transform step that causes issues when an underscore-prefixed
method is called outside of it's module, so changing the method names
seems like the easiest way to deal with this issue and still send the
message that they are private.

There is a way to turn off the transformation step on a file-by-file
basis but that would lead to touching more files than this approach, and
this code will be deprecated in the next release anyway.