Improved version of "Fix modifying state during onChange ... #667"
This fixes some issues with the fix originally merged from PR #667[2]

Credit to @balpert and @srmcconomy for this fix. For now I'm getting
them back in sync.

Also thanks to @davidchang for pinging me about landing this,
@colinjeane for reporting this bug in #917 and submitting a nearly
identical fix: https://github.com/facebook/draft-js/pull/919

@colinjeane since we already had this fixed internally I'm going to use
our approach to keep things in sync.

fixes #917, #849

[2]: https://github.com/facebook/draft-js/pull/667

We also had to tweak this compared to the current internal version to
accomodate a recent change that makes 'editOnBeforeInput' and the other
handlers take 'editor' as an explicit argument instead of assumming
'this' will be the editor;
https://github.com/facebook/draft-js/commit/e64c2c3f2a3654b925b18baff448100cfbbf3738

Note that we are updating the version of 'fbjs' required in order to get
access to the 'setImmediate' polyfill.