Outer-most block type should take preference
Currently unstyled blocks (but not their aliases) are ignored whenever any other blocks also need to be converted. This ensures that cases like `<div><h2>Hello world!</h2></div>` are properly associated marked as `header-two` but it creates a parsing difference between `<div>Hello</div><div>world</div>` and `<h2>Good morning!</h2><div>Hello</div><div>world</div>` in which the former has `Hello` and `world` in two different blocks but the latter has them in one.

The goal for parsing is that if the block that's being parsed is in something other than an unstyled block then it is part of its parent otherwise it's a new block. This ensures even unstyled blocks are always treated as blocks and moves the check for whether a block is unstyled deeper into the parsing logic.