Fake the input event for IE
When `_pendingStateFromBeforeInput` was added it defered some work in updating the editor's state until the next `input` event is received. Unfortunately, IE 11 doesn't support this event on contenteditable elements and React does not handle this inconsistency and so the defered work was never processed.

This change kicks off timer to fire as soon as we hand control back to the browser and so simulate the event occuring on IE.