Use position fixed for the paste trab
Position absolute will cause the paste trap to move around on the page if the editor itself is scrolled. Position fixed will keep it in the same position on the screen even if the editor is scrolled.

Ideally the paste trap would follow the location of the caret but that's expensive. For now, use fixed because we can assume that the caret will be on screen (or will become on screen) if we are typing in the editor. This prevents the browser from jumping around in the viewport as focus moves between the editor, paste trap, and back to the editor.