Validate the anchor and focusLeaf
This fixes textioHQ/frontend#2211

This is a continuation of other issues seen in this file, most likely
due to timing issues w.r.t the editorState and the DOM.

Basically, what happens is that this function is trying to take
selection information as from the DOM, and convert that into a logical
selectionState that matches whatever is in editorState. A previous fix
to this file validated the anchorBlockTree and focusBlockTree. This fix
continues the same pattern by validating that the editorState matches
the DOM.

There is this half-attempt already in the codebase to validate the
anchorLeaf & focusLeaf, but it is incomplete (and thus we are hitting
the crash trying to reference it below).