Update editor immediately for native insertion events (#24)
Fixes textioHQ/frontend#2259

This is an alternative method to getting around the problem of IE11 having no input event and wanting to allow native browser insertion to increase react performance
See also: facebook#667
See also: facebook#871

The premise of this is that it updates the editor state immediately, on the onBeforeInput event. And then it adds extra flags to know what to do during the render. As opposed to waiting for the input event on IE.