Fix tweet example for v0.11.0 syntax change
This is the one method signature that will change in a breaking way from
v0.10.1 to v0.11.0. We should aggressively call this out in the docs,
and maybe add some warning in the console or something.

There also seems to have been a typo that I fixed -
contentBlock.getPlainText -> contentBlock.getText