WIP Fixup diff for enabling 'ship-it' and 'import-it'
**WIP because:** Flow is not passing. We are on a later version of Flow
internally, which required a new 'FlowFixMe' comment. When I update the
Flow version in OS we get a new error about missing 'setImmediate'
module. Need to figure this out before landing.

**what is the change?:**
Development on Draft.js is causing Github and our internal version to
get out of sync.

This pulls various changes from our internal version into the Github
project, so that they will be basically the same.

Some of the included changes;
 - Reorder requires via a codemod, maintained by an ESLint rule
   (D5117037)
 - Fix 'first character prevention' bug that is triggered by Fiber async
   rendering (D5161053)
 - Add 'z-index' to placeholder (D4769133)
 - Add possible fix for issue #1020 (D5261172)
 - Add 'flowfixme' comment (D5337485)

For reference, this patch was generated by the commands listed in
D5273910.

**why make this change?:**
We want to keep them in sync going forward, and will be enabling scripts
to automate mirroring changes back and forth.

**test plan:**
`yarn install && yarn build && yarn test` and manually tested examples.

fbshipit-source-id:2b886ea7d631aefb30b440c0cd7a0e43f165791b