Fix minor build errors with 'universal' example
**what is the change?:**
Updates the path to find 'draft' from the example, since the file was
moved.

Also the '.gitkeep' in a directory wasn't there on this branch, and that
throws if you build for the first time and that directory hasn't been
created.

**why make this change?:**
To make it easy to build this example, for future testing.

**test plan:**
`cd examples/universal && yarn install && yarn demo`
Runs the 'universal' demo

**issue:**
https://github.com/facebook/draft-js/issues/835