Update handleKeyCommand examples and docs
handleKeyCommand receives a second parameter with the latest instance of the editor state. When this function needs the editor state it should use this instance instead of any other reference.

This update reflects a change made to handleKeyCommand to also receive the editor state. Since it can be potentially changed by Draft during key handling any previous references might be of an old version of the state.