Add test for rendering custom block with default block map
**what is the change?:**
Adding a test for the solution to issue https://github.com/facebook/draft-js/issues/1211

In general I prefer testing the public API, so we actually render an
example editor and trigger the situation which used to throw an error.

**why make this change?:**
We need better unit tests for Draft, and this will help stop the bug
from recurring if someone randomly removes that check in the future.

**test plan:**
`yarn run test src/component/contents/__tests__/DraftEditorContent.react-test.js`

We tried temporarily undoing the fix and running the test, and verified
that it failed.

**issue:**
https://github.com/facebook/draft-js/issues/1211