Tweak type check in DraftEditor '_focus' callback
**what is the change?:**
The problem - we currently throw if '_focus' receives an invalid type of
editorNode.

But in some cases the editorNode is null, and it's pretty
convenient/harmless to fail silently there.

So the proposed change was to fail silently if the editorNode was an
invalid type.

However, it's a bit stricter to fail only if the editorNode is undefined
or null, which is the case this is meant to address. If we get some
other type of thing as the editorNode somehow, then we still throw.

**why make this change?:**
To be slightly stricter and still solve the problem.

**test plan:**
Didn't really test it! :D

**issue:**
No issue open afaik