Exploration - Feature tree / Setting up BlockNode interface
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Setting up BlockNode interface**

This PR will be laying down the foundations by creating a tree block compatible node.

In order to avoiding including conditional logic as much as possible to existing draft library while exploring we will use interfaces to guarantee that the new tree block structure is some what compatible with the existing ContentBlock

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1477

Differential Revision: D6245392

fbshipit-source-id: 1cdc9d8de09f5174f1834020a78ad013da51a9d7