Exploration - Feature tree / Wiring up BlockNode interface
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Wiring up BlockNode interface**

This PR will make sure that ContentBlock references are now referencing the BlockNode interface in order for us to be able to support both ContentBlock and ContentBlockNode operations

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1493

Differential Revision: D6260642

fbshipit-source-id: 09f0cb2631579cbe9e9ac6fbdc1121f1d42d6b4e