Exploration - Feature tree / Serializing Draft state to tree data structure
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Serializing Draft state to tree data structure**

This PR will make sure that we can serialize existing draft state into a tree data block structure for ContentBlockNode and flat data structure for the existing ContentBlock

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1494

Differential Revision: D6267068

fbshipit-source-id: b120986d0caf49eda909a57fa21df42159508b64