Mark props that have defaults as required
Fixes facebook/draft-js#1514

According to the Flow documentation[0],
when using default props,
the prop should be marked as required,
as "Flow will make sure that [the prop]
is optional if you have a default prop for [it]."

This fixes Flow errors that would occur
when using Draft.js in a project with Flow.
Flow would complain that `blockRendererFn` and `blockStyleFn`
could potentially be `undefined`
when passed to `DraftEditorContents`,
even though these two properties have default values,
so they will never be undefined.

This commit also upgrades Flow to 0.59.0,
since there weren't any additional Flow errors
between the version of Flow that was being used (0.53.1)
and the latest version of Flow (0.59.0).

[0]: https://flow.org/en/docs/react/components/#toc-using-default-props