Exploration - Feature tree / Adding support for getting transaction fragments
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Adding support for getting transaction fragments**

This PR adds support for getting tree based fragments from the content state transaction

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1536

Differential Revision: D6448624

fbshipit-source-id: 17c321cc7f2eaed227de8e17e717aa2f91b3e571