Experiment - does updating Jest fix CI?
**what is the change?:**
Bump the Jest version, undoing the revert commit landed in
https://github.com/facebook/draft-js/commit/e127090a2c23ea4de3e7396fa310a3ecc8c735ce

**why make this change?:**
CI is failing to do a segfault, and we have no strong theories yet about
why. The last commit to land before it started failing was the revert
commit, so let's see if that's what caused the failure.

**test plan:**
push this to Github and see if CI passes

tests are passing locally

**issue:**