Restrict 'prettier' version to unbreak CI
**what is the change?:**
We have verified that upgrading to Prettier v1.9.0 or v1.9.1 both cause
a Segmentation Fault when running `eslint`. This happens in CI, and I
was able to verify it locally.

**why make this change?:**
We want to use Prettier in CI and not get segmentation faults.

**test plan:**
Locally I did -
`rm -rf node_modules && yarn install && yarn lint` and got no segfault.
Pushing this to github and waiting for CI to pass will further verify.

**issue:**
issue #1562