Fix typo in editOnBeforeInput
**what is the change?:**
We were calling `this.update` when there was no `update` method on
`this`. It should be `editor.update`.

This fix is not really complete in the first place, but I'll make any
follow-up changes in a new PR.

**why make this change?:**
This is causing an error to be thrown - see internal FB task t24443952

**test plan:**
I manually reproduce the bug before the fix -
(Flarnie will insert screenshot)

and then was not able to reproduce with the same steps after the fix.

To reproduce -
- open the 'plaintext' example and type 'hello'
- select the 'o' with a forwards selection, left to right
- type 'o'. This was throwing, now it's not.

**issue:**
internal task t24443952