fix getWhitespaceChunk
Summary:
This fixes `getWhitespaceChunk` to properly add the appropriate entity (if present). It seems that this was broken in [[ https://github.com/facebook/draft-js/pull/1518 | #1518 ]] (D6372235). The problem is that the `map` in
  new Array(1).map(none => (inEntity ? inEntity : none))
doesn't actually do anything since the array has no entries. To fix this, I'm just reverting the code to how it was previously.

Reviewed By: sophiebits

Differential Revision: D6669010

fbshipit-source-id: 17a430e36f75b48e7231619955dff73508f3fc2f