Add guard via early return
Need for guard is a result of the fact that the code now gets DOM
node’s owner window instead of global window. Previously, in the cases
where the DOM node isn’t in a window, the code would run but not do
anything. With this change, the function will early return and the rest
of the code won’t be run.