Adding aria-describedby attribute the draft editor
Summary:
We have a ariaDescribedBy prop in DraftEditorProps.js but it is not set in DraftEditor.
Currently the aria-describedby attribute is set based on if there is a placeholder and is set to the placeholderId.
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_aria-describedby_attribute

issue on github which also has suggested the fix: https://github.com/facebook/draft-js/issues/1214

Reviewed By: jessebeach, flarnie

Differential Revision: D6154932

fbshipit-source-id: d828ca5f73921c5d303820c5aed5b39eda08ccd1