Make test script run on Windows
The test and test-ci npm scripts currently rely on setting the NODE_ENV environment variable in a Unix-specific way. This moves the logic to set the environment variable into the cross-env package which makes this work on both Windows and Unix environments.

After applying this change I can successfully run `npm test` which I couldn't before.