Widen logging and add stack trace for IE selection bug
Summary:
We are hoping to remove this flag that was meant to fix an IE selection bug. It
blocks selection events from being handled while React is rendering - the thing
is, this has never been a problem that we can see.

We are logging to see if real users actually hit this bug and if the flag
prevents a bad state in that case.

If we find that the flag is never hit, then it will soon be removed, unlocking
removal of a deprecated lifecycle method.

Reviewed By: sophiebits

Differential Revision: D6971850

fbshipit-source-id: 8874b21161de3b3f0f568baca2a8c0278b01fa54