Exploration - Feature tree / Fixing range removal for adjcent retained parents
Fixing case where imediate parent is selected as anchor node. Due to how
we do moveSelectionForward and moveSelectionBackward we highjack the
natural selection and get in cases where the parent has the anchor or
focus where it does not have text.

This PR ensures that when the parent is selected and retained that its
previous block does not lose its links

***

**Note:**
This is unstable and not part of the public API and should not be used
by
production systems.