Fix broken links in Advanced-Topics-Entities.md
Apart from the proposed changes in this PR, several of permalinks in the documentation seem outdated and I reckon there might be other broken links pointing to `.html` extensions rather than `.md`. This is just a quick fix for some of the broken link I have found, but this area seems to require a bit more love.