Add utility methods for parent-child & sibling pointer updates
Summary:
When the tree is updated, we need to make sure that we correctly mirror the parent-child & prevSibling-nextSibling links. The utility methods in this diff help make these pointer updates more resilient.

This diff introduces methods to add:
- a child to a parent (either as the first or last child)
- a prev <-> next sibling relationship

These functions will only respect the local invariants on the nodes being updated, and can leave the rest of the tree in an inconsistent state that requires a few more updates (see the test file for examples of this).

In the next diff, I'll be codifying some common paradigms (such as creating a new parent node & adding an existing node as  its child) which will use the methods implemented in this diff to simplify the logic. The paradigm-based operations will preserve the tree invariant after the operation.

Differential Revision: D9601093

fbshipit-source-id: 3152ffc24a49e7a591958aada3333f331aaef9dd