Update Advanced-Topics-Entities.md (#1767)
Summary:
Add a line of code in the example to set the editor state after adding the entity.

For a newcomer, it's not straightforward because it wasn't mentioned anywhere in the docs prior to arriving here. Of course, you'll figure it out once you check the example. But let's save newcomers a roundtrip

**Summary**

Update the docs
Pull Request resolved: https://github.com/facebook/draft-js/pull/1767

Differential Revision: D9749305

fbshipit-source-id: 88af5ee30690fa41cc72280a6bad153a17324c33