Allow Option+Space to be handled on OSX Chrome
Currently pressing Option+Space in Chrome on OSX will cause a non breaking space to always be inserted into the editor. This behavior ignores any key bindings that the editor has set.

This change moves the logic which adds the non breaking space to after the point that the key binding function has been run in order to give editors a chance to handle this behavior on their own.