Styling Controls in UI Readme
Summary:
The current example for styling controls (applying bold, italics etc.) using UI causes the editor to lose focus. This can be resolved by using the onMouseDown event to toggleInlineStyle. This event does not cause the editor to lose focus. Any character typed after that is "boldened" which is what we want.