Add rules to .flowconfig for flow strict rules (#2288)
Summary:
Internally at facebook, we use `flow strict` to define certain rules that make our codebase more reliable. I just noticed the .flowconfig for this repo doesn't include rules for `flow strict`. We do have files in the repo annotated with strict flow, so lets configure it to get the benefits.

Unfortunatley, there's 3 rules we have enabled internally which we can't use in the repo as it exists today. I put the necessary info as a comment on the .flowconfig itself.

npm run flow -> no errors.
Pull Request resolved: https://github.com/facebook/draft-js/pull/2288

Differential Revision: D19265177

Pulled By: mrkev

fbshipit-source-id: 0eca719027d62c2681cd09be03385ba53a27c2e3