Fix iOS double-space quick-type bug
Summary:
Using a Pull Request from github that was made to fix this very bug, I was able to patch the code and fix the problem for us in the meantime.

Since the open-source and internal codebases are totally out of sync, it's not easy to merge the original PR right now.

https://github.com/facebook/draft-js/pull/1152

Reviewed By: claudiopro

Differential Revision: D20016281

fbshipit-source-id: aeefd5c3f17c51c62dd5d7c225cf62d0cd2db26f