Update yarn.lock to fix travis build. (#2355)
Summary:
Travis is failing with `The command "eval yarn --frozen-lockfile " failed.` The lockfile is out of sync with package.json, which is usually not fatal but the ` --frozen-lockfile` flag means this throws.

This change fixes it, by doing `yarn install` and committing the changed lockfile.

------------------
EDIT: Also changes the flow name mapping for ReactDOM. It was failing since it's now ReactDOMComet.
Pull Request resolved: https://github.com/facebook/draft-js/pull/2355

Reviewed By: claudiopro

Differential Revision: D20532615

Pulled By: mrkev

fbshipit-source-id: 666399be23772fc7383ced8875bb0220a4abedd2