fix pasting invalid links
Summary:
I got a bug report today where someone was trying to copy/paste some content and nothing happened at all. I discovered that this was due to a bad link in her content, something like this:

```
<a href="https:// https://www.facebook.com">Faceboook</a>
```

In Chrome, this passes all the checks in `isValidAnchor`, but throws an error later on this line inside `_addAnchorNode`:
```
entityConfig.url = new URI(anchor.href).toString();
```
To account for this, I added an explicit `URI.tryParseURI` inside `isValidAnchor`.

Reviewed By: claudiopro

Differential Revision: D22601322

fbshipit-source-id: 61cf050bb6e4a31bbe95995ad3f57e8f39a1d954