Remove paste handling logic for images with "presentation" role
Summary: This is essentially a revert of D9333247 (https://github.com/facebook/draft-js/commit/6df3808c6d0453a68cd6f9939fec80a527cca497) which I worked on a couple of years ago. The gk `draftjs_fix_paste_for_img` was never opened to public. I am deleting this code because I feel its not an ideal solution and the issue is a very minor cosmetic issue that oc curs only in the specific situation of pasting a text into draft with an fbemoji. I will clean up the gk after this diff lands.

Differential Revision: D23167522

fbshipit-source-id: 7f9252eda4786bbb17c7bf25939eade937d28689