Revise `onKeyDown` logic for handling composition resolution
Summary: Looking through the code, it seems that there is a missing condition during the `onKeyDown` handling for composition. Surely, we should be also checking that the resolved flag is not set, otherwise we might over-fire `resolveComposition`? This might possibly be one of the causes for the IME issues we're seeing.

Differential Revision: D23734060

fbshipit-source-id: f94d58851019abe1ab7dc1b15622219b164f2293