Fix composition error when using IME characters
Summary: This diff fixes an issue that we are tracking to do with Japanese/Korean IME character input and the mentions type ahead. Specifically, it's when we're resolving composition, we don't want the mentions strategy to change the editor state, which results in the selection cursor to incorrectly be at a position that isn't actually possible.

Differential Revision: D23840422

fbshipit-source-id: ef8cc9056b9e5943671c13a35a55454cfe9c0f9d