convertFromHTMLToContentBlocks
Summary:
I'm removing usage of `DraftEntity` directly, since this is an old API that has been deprecated for at this point years, and is blocking the fixing of some bugs.

Instead, entities should be accessed through the content state for an editor.

This function is very similar to `EditorCoreHTMLToContentBlocks`, and the change is essentially the same as the one in D24036853.

Reviewed By: AjkP

Differential Revision: D24036913

fbshipit-source-id: aee294b82e103959726524f80f2d4a9e1e41df70